requirements.txt
pymongo==3.12.1
elasticsearch==7.15.1
requests==2.27.1
python-dotenv==0.20.0